Late commit: adaptation of DT-XAI (Trustee-based) to streaming anomaly detection. Details: - First round (reproducing batch DT-XAI in my SIURU pipeline) successful, but newer changes apparently broke that. - Second round (executing streaming DT-XAI) was completed only up to the training and testing of River models, but I need to modify Trustee to receive River objects and execute a loop of the method predict_one (River) reading the entire traffic dataset sample by sample, instead of a single pass of the method predict (scikit-learn) with the entire dataset.
OBS1: I don't remember very well, but I fixed a problem in Hung's adaptation
of SIURU to apply Trustee. He adapted only the MultiSampleEncoder, but when
I tried to use the DefaultEncoder (single sample), which is the only
one possible to use with River, it did not work. I fixed something involving
the encoders, the main program and the RandomForestXAI method, and then I
eproduced batch DT-XAI in my SIURU pipeline (but note that the main prorgam
IoT-AD-xai.py is a simple copy of Hung's version with small changes for my
pipeline).
OBS2: unfortunately, I did not register the changes made in git.
OBS3: River ML algorithms do not have predict methods, only predict_one, so
they cannot use the MultiSampleEncoder.