Streaming DT-XAI working! Trustee is compatible with River now. It was an incompatibility bug, not a problem in the data augmentation. Description of the problems below:
It was not a problem in data augmentation as I thought. Instead, it was
an bug related to the incompatibility between scikit-learn and River.

- First and foremost, the type of data required by River is a dict with
keys. For some reason I thought that these keys did not need to be strictly
the same strings, only the same quantity, because I remembered doing this
before inside SIURU (I probably remember wrongly). Fixing this was the
main issue.
- But there was a problem of my SIURU ML class (e.g. AdaptativeRandomForestModelXAI)
overriding the method get_model_for_trustee(self) from the parent class
ITrusteeExplainableModel. The later returns self (which is AdaptativeRandomForestModelXAI)
but the former returns self.model_instance, which is a River class
(ARFClassifier). This mismatch of class type confused the heck out of
me and added to the problem of scikit ML classes having the
predict method (which SIURU's scikit-based ML classes do have),
while River ML classes have only the predict_one method (and
SIURU's River-based ML classes follow the scikit-based ones: only
predict method).
- I need to double check, but I train the SIURU's River-based class
with a scaler, transforming model_instance in a River pipeline of
(scaler | ARFClassifier). On the other hand, the River ML caller in
Trustee (which has predict_one instead of predict), calls predict_one
without scaling first (but I really need to double check, because
printing the type of the object returns a pipeline).